pyproject.toml
[project]
name = "video-api"
version = "0.1.0"
description = ""
authors = [
    {name = "gagarin yaroslav",email = "gagarinyaroslav3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
test = [
    "pytest (>=9.0.2,<10.0.0)"
]
